All termination methods work properly now (except timer taking too long)

+++ master.c
@@ -23,19 +23,7 @@ int shmAllocated;
 int shmid;
 char *programName;
 int activeProcesses;
-
-
-/* Function Prototypes */
-int isANumber(char*);
-char *getOutputPerror();
-int deallocateSharedMemory();
-void endProgramHandler(int);
-void childTermHandler(int);
-void ctrlCHandler(int);
-void logTermination(char*);
-static void timeoutHandler(int);
-static int setupinterrupt(void);
-static int setupitimer(void);
+int currentlyTerminating;
 struct shmseg *shmp;
 
 struct shmseg {
@@ -44,19 +32,38 @@ struct shmseg {
 	int choosing[MAX_PROCS];
 };
 
+
+/* Function Prototypes */
+int isANumber(char*);
+char *getOutputPerror();
+int deallocateSharedMemory();
+void endProgramHandler(int, int);
+void childTermHandler(int);
+void ctrlCHandler(int);
+void logTermination(char*);
+static void timeoutHandler(int);
+static int setupinterrupt(void);
+static int setupitimer(void);
+
+
 int main (int argc, char *argv[]) {
+	/* Setup signals/interrupts */
 	signal(SIGINT, ctrlCHandler);
 	signal(SIGCHLD, childTermHandler);
+	
+	/* Initialize variables */
 	programName = argv[0];
 	shmAllocated = 0;
 	activeProcesses = 0;
 	ss = 100;
 	nprocs = 0;
+	currentlyTerminating = 0;
 	
 	int i;
 	int usageStatement = 0;
 	int option;
 	while ((option = getopt(argc, argv, "ht:")) != -1) {
+		/* Parse the options -h and -t */
 		switch (option) {
 			case 'h' :
 				usageStatement = 1;
@@ -75,8 +82,10 @@ int main (int argc, char *argv[]) {
 		}
 	}
 	if (argv[optind] == NULL) {
+		/* User did not enter a value for n */
 		usageStatement = 1;
 	} else {
+		/* Parse the argument for n */
 		if (!isANumber(argv[optind])) usageStatement = 1;
 		else {
 			if (atoi(argv[optind]) > MAX_PROCS) {
@@ -89,6 +98,8 @@ int main (int argc, char *argv[]) {
 			}
 		}
 	}
+	
+	/* Print usage statement if the user used -h or entered an argument improperly */
 	if (usageStatement) {
 		printf("Usage: ./master [-h] [-t ss] [n]\n");
 		printf("Runs master program to fork 'n' slave processes.\n");
@@ -99,9 +110,8 @@ int main (int argc, char *argv[]) {
 		return 0;
 	}	
 	
-	shmAllocated = 1;
-	
 	/* Allocate shared memory */
+	shmAllocated = 1;
 	shmid = shmget(SHM_KEY, sizeof(struct shmseg), SHM_PERM|IPC_CREAT);
 	if (shmid == -1) {
 		char *output = getOutputPerror();
@@ -165,6 +175,7 @@ int main (int argc, char *argv[]) {
         return 0;
 }
 
+/* Returns false/true (0/1) value if the string is an integer */
 int isANumber (char *str) {
 	int i;
 	for (i = 0; i < strlen(str); i++) {
@@ -173,15 +184,16 @@ int isANumber (char *str) {
 	return 1;
 }
 
+/* Return the beginning of the error message for perror */
 char *getOutputPerror () {
 	char* output = strdup(programName);
 	strcat(output, ": Error");
 	return output;
 }
 
+/* Removes shared memory attachment and deletes/deallocates the shared memory segment */
 int deallocateSharedMemory() {
 	int returnValue;
-	printf("Deallocating shared memory...\n");
 	returnValue = shmdt(shmp);
 	if (returnValue == -1) {
 		char *output = getOutputPerror();
@@ -199,26 +211,32 @@ int deallocateSharedMemory() {
 	return 0;
 }
 
-void endProgramHandler (int s) {
-	int i;
-	for (i = 0; i < nprocs; i++) {
-		if ((kill(children[i], SIGKILL)) == -1) {
-			char *output = getOutputPerror();
-			perror(output);
+/* Function to end program. Kills all children and calls function to deallocate shared memory */
+void endProgramHandler (int s, int killChildren) {
+	if (killChildren) {
+		int i;
+		for (i = 0; i < nprocs; i++) {
+			if ((kill(children[i], SIGKILL)) == -1) {
+				char *output = getOutputPerror();
+				perror(output);
+			}
 		}
 	}
 	if (shmAllocated) deallocateSharedMemory();
 	exit(0);
 }
 
+/* For timer */
 static int setupitimer(void) {
 	struct itimerval value;
+	printf("ss == %d\n", ss);
 	value.it_interval.tv_sec = ss;
 	value.it_interval.tv_usec = 0;
 	value.it_value = value.it_interval;
 	return (setitimer(ITIMER_PROF, &value, NULL));
 }
 
+/* For timer */
 static int setupinterrupt(void) {
 	struct sigaction act;
 	act.sa_handler = timeoutHandler;
@@ -226,26 +244,33 @@ static int setupinterrupt(void) {
 	return (sigemptyset(&act.sa_mask) || sigaction(SIGPROF, &act, NULL));
 }
 
+/* Detected ctrl+c, so we can end the program */
 void ctrlCHandler(int s) {
+	printf("\n Detected ctrl+c. Now exiting program...\n");
+	currentlyTerminating = 1;
 	logTermination("ctrl+C");
-	endProgramHandler(1);
+	endProgramHandler(1, 1);
 }
 
+/* The timer has ended, so we can end the program */
 static void timeoutHandler(int s) {
+	currentlyTerminating = 1;
 	printf("Timer ended. Now exiting program...\n");
 	logTermination("timeout");
-	endProgramHandler(1);
+	endProgramHandler(1, 1);
 }
 
+/* All children have terminated, so we can end the program */
 void childTermHandler(int s) {
 	activeProcesses--;
-	if (activeProcesses < 1) {
+	if (activeProcesses < 1 && !currentlyTerminating) {
 		printf("All children have terminated. Now exiting program...\n");
 		logTermination("all children terminated");
-		endProgramHandler(1);
-	}		
+		endProgramHandler(1, 0);
+	}
 }
 
+/* Print message to 'cstest' about program termination */
 void logTermination(char *method) {
 	time_t rawtime;
 	struct tm * timeinfo;
